Adding pictures with shareable, restrict routes
server.js Changes:
*Changed tabs to 4 spaces
*Fixed indenting to be consistent with other functions. I didn't fix all
of the indenting, stopped at a certain point.
*Adding the restrict middleware function to all routes requiring login.
*Removed the /images post route.
*Enabled images to be posted along with a shareable. the field name for
the image is picture, and the name of the picture is stored in the
database.
*Multer middleware is used in the makeshareablerequest and
makeshareableoffer routes.

Database Changes:
*Users table now has a points field, you get points for doing things in
the app.
*Users table no longer has a rank_id field since the points will
determine their rank.
*Ranks now have a threshold, your rank corresponds to the greatest
threshold that you surpass,
ex: rank master has threshold 8, when you get 8 points you become a
master.
*Shareable table now has shar_pic_name field for the image name that is
stored for a shareable. This can be null.

sql_queries.js Changes:
*Changed addshareable so that it now adds the shar_pic_name field.

utils.js Changes:
*Added a function isEmptyObject. This should be a function included in
nodeJS but it isn't. I used it for testing and debugging for
implementing the image uploads.