conn -> var conn &&& throw(err) -> fn(err)
*in the /offers route changed the json object "usersOffers" to "offers"
this is to match the route.

*in the /requests route changed the json object "usersRequests" to
"requests" this is to match the route.

In sql_queries.js changed all declarations of the mysqlconnection object
"conn" to "var conn" This way the connection variable is not global.

Same thing mentioned above done for queryString variable.

Changed all functions with callbacks to not throw errors but to instead
pass the errors to the callback functions so they can be handled there.