Changed the order of middleware, body parser is last
Cookieparser should be first, then session handler, then body parser.

The bodyparser was giving deprecated warnings because of the way we were
initializing the middleware, I fixed that.

I also added middleware that gives an info log when each one of these
middlewares completes in order to nab the issue that is still causing
random hang ups in the web server. Activate this by setting debugHangup
to TRUE. It is deactivated by default so your terminal doesn't get
spammed.

Also fixed a string which used "request" instead of "offer"