viewreqoffshareable change
Has to return object with property "transactions" if you are the creator
of the shareable.
Has to return object with property "transaction" if you are not the
creater.

This was broken in commit 57a460a117cb9faea8515a66ce13b2cbe9477e2c
because it was always returning transactions if trnsction array was
empty.